Full implementation of the missile
Missile ability has now been properly added. Now when the player selects the ability cards the missile will receive the described changes.

Also fixed multiple bugs with the missile object. Mainly focusing on enemies being hit by the same object multiple times.

Finally, reduced the size of the missile hit markers as they can really take up the screen if they hit a big group of enemies.

+++ Assets/Scripts/Abilities/Missle.cs
@@ -12,7 +12,7 @@ public class Missle : MonoBehaviour
     private float angle = 0f;
     private Vector2 previousPosition;
     public Queue<GameObject> enemyQueue;
-    public int missleNumber;
+    public float missleNumber;
 
     public GameObject missle;
     public float spawnTimer;
